requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-docx>=0.8.11
python-dotenv>=1.0.0
jinja2>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
setuptools>=65.0.0